Refactor filter components
